code refactoring, class inheritance added for input file types (txt and json)
