--- module-8-python-basics/json_input.json
@@ -1,31 +1,0 @@
-{
-	"data": [
-		{
-			"type": "News",
-			"content": {
-				"news": "news from json.",
-				"city": "city"
-			}
-		},
-		{
-			"type": "News",
-			"content": {
-				"news": "news from json!",
-				"city": "city"
-			}
-		},
-		{
-			"type": "Ad",
-			"content": {
-				"ad": "adverTising from json.",
-				"date": "2/2/2022"
-			}
-		},
-		{
-			"type": "Movie",
-			"content": {
-				"movie": "movie From json."
-			}
-		}
-	]
-}
