+++ module-8-python-basics/letter-statistic.csv
@@ -1,22 +1,24 @@
 letter,count_all,count_uppercase,percentage
-n,11,4,8.94
-e,11,1,8.94
-w,4,0,3.25
-s,11,0,8.94
-o,12,0,9.76
-j,4,0,3.25
-f,6,0,4.88
-r,6,0,4.88
-m,7,2,5.69
-y,4,0,3.25
-i,10,0,8.13
-t,10,0,8.13
-c,3,2,2.44
-d,4,1,3.25
-p,1,1,0.81
-a,8,3,6.5
-v,4,0,3.25
-g,1,0,0.81
-l,3,0,2.44
-u,2,0,1.63
-h,1,0,0.81
+w,10,0,1.95
+s,40,6,7.81
+n,48,10,9.38
+e,55,3,10.74
+i,44,1,8.59
+d,23,3,4.49
+p,12,4,2.34
+t,39,1,7.62
+h,14,0,2.73
+r,24,0,4.69
+o,39,0,7.62
+a,42,12,8.2
+g,19,3,3.71
+m,13,4,2.54
+l,24,4,4.69
+v,12,0,2.34
+k,6,0,1.17
+b,2,0,0.39
+u,11,0,2.15
+c,6,0,1.17
+f,16,0,3.12
+y,10,1,1.95
+j,3,0,0.59
