+++ module-8-python-basics/newsfeed.txt
@@ -1,20 +1,60 @@
 News ------------------------
-News from json.
-City, 17/10/2021 17.28
------------------------------
-
-News ------------------------
-News from json!
-City, 17/10/2021 17.28
+News 1. Something happend. It's interesting.
+London, 17/10/2021 22.42
 -----------------------------
 
 Private Ad ------------------
-Advertising from json.
-Actual until: 2/2/2022, 107 days left
+Advertising 1. Sell bike. 500 eur.
+Actual until: 12/11/2021, 25 days left
+-----------------------------
+
+News ------------------------
+News 2. Something happend again!
+Grodno, 17/10/2021 22.42
 -----------------------------
 
 Movie of the Day --------------
-Movie from json.
-17/10/2021 17.28. Estimation: 5
+The avengers
+17/10/2021 22.42. Estimation: 6
+------------------------------
+
+Movie of the Day --------------
+Sherlok
+17/10/2021 22.42. Estimation: 4
+------------------------------
+
+Private Ad ------------------
+Advertising 3. Looking for python courses.
+Actual until: 6/6/2022, 231 days left
+-----------------------------
+
+Private Ad ------------------
+Advertising 4. Looking for python courses.
+Actual until: 6/6/2022, 231 days left
+-----------------------------
+
+News ------------------------
+News 3. Something happend again!
+Grodno, 17/10/2021 22.42
+-----------------------------
+
+News ------------------------
+News from json file!
+Minsk, 17/10/2021 22.43
+-----------------------------
+
+News ------------------------
+News from json file again? Yes!
+Grodno, 17/10/2021 22.43
+-----------------------------
+
+Private Ad ------------------
+Advertising from json file. Sell bike again.
+Actual until: 9/2/2022, 114 days left
+-----------------------------
+
+Movie of the Day --------------
+Lord of the rings
+17/10/2021 22.43. Estimation: 6
 ------------------------------
