+++ module-8-python-basics/word-count.csv
@@ -1,16 +1,37 @@
-news-4
-of-1
-ad-1
-private-1
-actual-1
-day-1
-left-1
-until-1
-the-1
-city-2
-estimation-1
-from-4
-json-4
-days-1
-advertising-1
-movie-2
+private-4
+from-3
+again-4
+estimation-3
+python-2
+happend-3
+sell-2
+rings-1
+actual-4
+movie-3
+day-3
+for-2
+news-10
+eur-1
+it-1
+until-4
+sherlok-1
+grodno-3
+days-4
+left-4
+yes-1
+is-1
+interesting-1
+the-5
+advertising-4
+london-1
+courses-2
+something-3
+bike-2
+minsk-1
+file-3
+avengers-1
+ad-4
+of-4
+json-3
+lord-1
+looking-2
